Chore: rename module cli_parse to cli_in
